refactor(terminal): split ContentGrid render branches and tip data
Extract the 1504-line ContentGrid into a custom hook (useContentGridContext),
5 branch subcomponents (fleet scope, maximized group, maximized single,
two-pane split, default grid), and 2 utility files (contentGridTips,
ContentGridEmptyState). GridShell provides shared ContextMenu +
TerminalCountWarning scaffolding for the 3 branches that need it.

- No behavior changes — every effect ordering, dependency array, ref
  attachment, and closure identity is preserved.
- Each extracted component carries its own "use memo" directive for
  React Compiler optimization across file boundaries.
- combinedGridRef stays useCallback-memoized with [setNodeRef] dep to
  prevent React 19 dnd-kit infinite re-render.
- 3 preMaximizeLayout cleanup effects preserve original ordering sharing
  preMaximizeLayoutRef.
- ResizeObserver dependency array unchanged.
- All 35 existing tests pass with path-only updates.
