+++ src/components/Terminal/__tests__/contentGridEmptyState.test.ts
@@ -2,23 +2,24 @@ import { describe, it, expect } from "vitest";
 import { readFile } from "fs/promises";
 import { resolve } from "path";
 
-const GRID_PATH = resolve(__dirname, "../ContentGrid.tsx");
+const EMPTY_STATE_PATH = resolve(__dirname, "../ContentGridEmptyState.tsx");
+const TIPS_PATH = resolve(__dirname, "../contentGridTips.tsx");
 
 describe("ContentGrid EmptyState — RecipeRunner integration", () => {
   it("hero section uses reduced spacing (mb-6 / mb-4)", async () => {
-    const content = await readFile(GRID_PATH, "utf-8");
+    const content = await readFile(EMPTY_STATE_PATH, "utf-8");
     expect(content).toContain('"mb-6 flex flex-col items-center text-center"');
     expect(content).toContain('"relative group mb-4"');
   });
 
   it("tip text uses /70 opacity, not /60", async () => {
-    const content = await readFile(GRID_PATH, "utf-8");
+    const content = await readFile(TIPS_PATH, "utf-8");
     expect(content).toContain("text-daintree-text/70 text-center");
     expect(content).not.toContain("text-daintree-text/60 text-center");
   });
 
   it("renders RecipeRunner component instead of inline recipe list", async () => {
-    const content = await readFile(GRID_PATH, "utf-8");
+    const content = await readFile(EMPTY_STATE_PATH, "utf-8");
     expect(content).toContain("<RecipeRunner");
     expect(content).toContain('from "./RecipeRunner/RecipeRunner"');
     // No inline recipe list markup should remain
@@ -29,17 +30,17 @@ describe("ContentGrid EmptyState — RecipeRunner integration", () => {
 
 describe("ContentGrid TIPS live shortcut migration — issue #6437", () => {
   it("imports useKeybindingDisplay so tip shortcuts react to remaps", async () => {
-    const content = await readFile(GRID_PATH, "utf-8");
+    const content = await readFile(TIPS_PATH, "utf-8");
     expect(content).toMatch(/import \{ useKeybindingDisplay \} from "@\/hooks\/useKeybinding"/);
   });
 
   it("declares a messageWithShortcut field on TipEntry", async () => {
-    const content = await readFile(GRID_PATH, "utf-8");
+    const content = await readFile(TIPS_PATH, "utf-8");
     expect(content).toMatch(/messageWithShortcut\?: \(shortcut: string\) => React\.ReactNode/);
   });
 
   it("defines a LiveTipMessage component that calls useKeybindingDisplay", async () => {
-    const content = await readFile(GRID_PATH, "utf-8");
+    const content = await readFile(TIPS_PATH, "utf-8");
     expect(content).toContain("function LiveTipMessage");
     const liveTipMatch = content.match(/function LiveTipMessage\([\s\S]*?\n\}/);
     expect(liveTipMatch).not.toBeNull();
@@ -48,29 +49,20 @@ describe("ContentGrid TIPS live shortcut migration — issue #6437", () => {
   });
 
   it("falls back to the static tip.message when no shortcut is bound", async () => {
-    const content = await readFile(GRID_PATH, "utf-8");
+    const content = await readFile(TIPS_PATH, "utf-8");
     const liveTipMatch = content.match(/function LiveTipMessage\([\s\S]*?\n\}/);
     expect(liveTipMatch).not.toBeNull();
-    // The component must render `tip.message` when shortcut is empty so unbound
-    // tips degrade gracefully instead of rendering "Press  to ..." with a blank
-    // <Kbd>.
     expect(liveTipMatch![0]).toMatch(/return <>\{tip\.message\}<\/>/);
   });
 
   it("RotatingTip renders <LiveTipMessage tip={tip} /> instead of {tip.message} directly", async () => {
-    const content = await readFile(GRID_PATH, "utf-8");
+    const content = await readFile(TIPS_PATH, "utf-8");
     expect(content).toContain("<LiveTipMessage tip={tip} />");
-    // The previous direct-render pattern should be gone from the RotatingTip JSX
-    // (the static `message` field is still referenced inside LiveTipMessage as
-    // the fallback path, which is fine).
     expect(content).not.toMatch(/Tip: \{tip\.message\}/);
   });
 
   it("shortcut-driven tips supply messageWithShortcut so the live combo renders inline", async () => {
-    const content = await readFile(GRID_PATH, "utf-8");
-    // Spot-check a few representative keybinding-driven tips. Each should
-    // declare a messageWithShortcut function so the live combo replaces the
-    // hardcoded glyph when the action is bound.
+    const content = await readFile(TIPS_PATH, "utf-8");
     const tipIds = [
       "quick-switcher",
       "panel-palette",
@@ -86,10 +78,7 @@ describe("ContentGrid TIPS live shortcut migration — issue #6437", () => {
   });
 
   it("worktree-overview tip uses worktree.overview (the toggle action) for shortcut display", async () => {
-    const content = await readFile(GRID_PATH, "utf-8");
-    // The toggle is what owns the Cmd+Shift+O default binding; the .open
-    // variant has no binding registered, so we must look up the toggle to get
-    // a non-empty display combo.
+    const content = await readFile(TIPS_PATH, "utf-8");
     const tipBlock = content.match(/id: "worktree-overview",[\s\S]*?(?=\n {2}\},)/);
     expect(tipBlock).not.toBeNull();
     expect(tipBlock![0]).toContain('shortcutActionId: "worktree.overview"');
